[luci] Reduce duplicate codes in weights quantization
This reduces duplicate codes in weights quantization.

ONE-DCO-1.0-Signed-off-by: Hyukjin Jeong <hj1.jeong@samsung.com>
